requirements.txt
pipdnslib==0.9.23
dnspython==2.4.2
python-nmap==0.7.1
